## Whats added: - Split topBar logic - Merge duplicate code in topBAr in one ## Issue(#1527)
