+++ save-frontend/src/test/kotlin/com/saveourtool/save/frontend/components/TopBarTest.kt
@@ -1,5 +1,6 @@
 package com.saveourtool.save.frontend.components
 
+import com.saveourtool.save.frontend.components.topbar.topBar
 import com.saveourtool.save.frontend.externals.*
 import com.saveourtool.save.info.UserInfo
